feat: connect to database
Implement a connection to the database, and an API `getLatestBlock`
to retrieve data from the database.

Implement a (initial?) module structure for `database` and
`MasterChain`.

When the API is used, currently an error is thrown saying that
retrieving the latest block failed (which is correct because there
is no data yet).