fix: solve errors with importing dependencies
The Partisia modules could not be imported because they are ESModules.
The project was configured for CommonJS.

Changed the configuration. As a side-effect all references to local
`.ts`-files has to get extension `.js`.

There are still problems with using the Command Line.
`commander` is now used instead of `@nest-command`,
but using `` results in an error.

Not a big issue, the command line is only used for testing.